+++ source/extensions/load_balancing_policies/common/load_balancer_impl.h
@@ -425,7 +425,7 @@ private:
 
   // Config for zone aware routing.
   const uint64_t min_cluster_size_;
-  const uint64_t force_local_zone_min_size_;
+  const uint32_t force_local_zone_min_size_;
   // Keep small members (bools and enums) at the end of class, to reduce alignment overhead.
   const uint32_t routing_enabled_;
   const bool fail_traffic_on_panic_ : 1;
